fix: wrong module id(name) on UpsertPolicy
We have been experimenting with delta bundles, and realized they are not working properly
with v0.42.x and 0.43, where you get you get this message whenever any policies contain a default rule:

`rego_type_error: multiple default rules named foo found`

When the bundle server receives a delta bundle there is a mismatch between the policy names(id?) in
`compiler.Modules` and `extraModules` even though the policies are identical.

This causes the `remainingAndExtra` variable in
`activateBundles` to contain one policies with the key `bundle/policy.rego` and one key with `/bundle/policy.rego`
even though they have the same policy.

I'm not completely sure if I have understood correctly, but this seems to be an issue because even though they have
different names, the path will end up being the same after trimming away `/` in the path.

This PR is a quick fix to ensure the name of policy remains consistent, not starting with a `/`.

We noticed that `ID` and `Name` have been used interchangingly on the module types, so shout out if we should change it.

Signed-off-by: Martin Johansen <martin.nordby.johansen@autodesk.com>